requirements/_base.txt
dj-database-url==0.4.2
Django==2.0.9
django-mptt==0.9.1
django-widget-tweaks==1.4.1
python-decouple==3.1
pytz==2018.5
Pillow==5.1.0
wordfilter==0.2.6.2
django-celery==3.2.2
django-celery-email<2.0.0
mistune==0.8.3
django-ipware==2.1.0
martor==1.3.3
Markdown==2.6.11
requests==2.20.0
django-markdownify==0.8.0
django-tinymce-4==0.0.23
django-tinymce4-lite==1.7.2
django-notifications-hq==1.5.0
beautifulsoup4==4.7.1
django-gravatar2==1.4.2
django-tenant-schemas==1.9.0